fixed code

+++ js/script.js
@@ -12,7 +12,7 @@ function calculate () {
   // input
   const length = parseFloat(document.getElementById("length-of-pyramid").value)
   const width = parseFloat(document.getElementById("width-of-pyramid").value)
-  const height = parseInt(document.getElementById('height-of-pyramid').value)
+  const height = parseFloat(document.getElementById('height-of-pyramid').value)
 
   // process
   const volume = (length * width * height) / 3
